[Move #234] 절대경로로 변경
