+++ app/(withGlobalLayout)/main/page.tsx
@@ -16,7 +16,7 @@ import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 import { toast } from "react-toastify";
-import MainCreateRoom from "../../../components/main/MainCreateRoom";
+import MainCreateRoom from "@/components/modal/CreateRoomModal";
 import { useCreateStore } from "../../../store/toggle-store";
 
 const Mainpage = () => {
